FERRY DONE

+++ src/ad222kr_assign2/e_2/Car.java
@@ -5,8 +5,8 @@ package ad222kr_assign2.e_2;
  */
 public class Car extends Vehicle {
   private final static int SPACE_REQUIRED = 5;
-  private final static double PASSENGER_COST = 15.0;
-  private final static double VEHICLE_COST = 100.0;
+  private final static int PASSENGER_COST = 15;
+  private final static int VEHICLE_COST = 100;
   private final static int MAX_NUMBER_OF_PASSENGERS = 4;
 
   public Car(int passengerCount) {
